users put and list
user crud put and list actions
fixing json output not to use a data attribute in json
fixed all tests not to depend on data attribute
protected api action- requires login

+++ src/Application/CRUD/CRUDObject.php
@@ -8,6 +8,10 @@ use App\CoreComponents\DatabaseConnection;
 use Exception;
 use App\CoreComponents\DatabaseInterface;
 
+/**
+ * A mini ORM class for CRUD objects. Controls CRUD operations of its descendants automatically based on setup
+ * protected attributes.
+ */
 abstract class CRUDObject {
 
     protected $_TABLENAME = '';
@@ -43,19 +47,18 @@ abstract class CRUDObject {
      */
     protected array $_validators = array(
     );
-    
     protected DatabaseInterface $_database;
     protected array $_data = array();
     protected array $_tableKeyAttribute = array(); //attributeName,fieldName,type
     protected array $_attributeTypes = array();
     protected array $_readOnlyAttributes = array();
-    protected array $_setterPreprocessors= array();
-    protected array $_attributeToField=array();
-        
+    protected array $_setterPreprocessors = array();
+    protected array $_attributeToField = array();
+
     public function __construct($databaseKey = null) {
 
-        $this->_database= DatabaseConnection::getDatabase();
-        
+        $this->_database = DatabaseConnection::getDatabase();
+
         reset($this->_dataFields);
         foreach ($this->_dataFields as $fieldName => $attributeData) {
             if (!empty($attributeData['isPrimary'])) {
@@ -64,11 +67,11 @@ abstract class CRUDObject {
 
             $this->_attributeTypes[$attributeData['attribute']] = $attributeData['type'];
             $this->_attributeToField[$attributeData['attribute']] = $fieldName;
-            
+
             if (!empty($attributeData['canNotChange'])) {
                 $this->_readOnlyAttributes[$attributeData['attribute']] = 1;
             }
-            
+
             if (!empty($attributeData['setterPreprocessor'])) {
                 $this->_setterPreprocessors[$attributeData['attribute']] = $attributeData['setterPreprocessor'];
             }
@@ -80,23 +83,42 @@ abstract class CRUDObject {
             $this->read($databaseKey);
         }
     }
-    
-    public function getTableName() {
+
+    /**
+     * Returns the object table name
+     * @return string
+     */
+    public function getTableName(): string {
         return $this->_TABLENAME;
     }
-    
-    public function getDeletedWhere() {
-        return $this->_REALDELETE?'':sprintf(' and %1$s=0 ', $this->_DELETEDFIELDNAME);
+
+    /**
+     * Returns an extra statement fof filtering soft deleted records
+     * @param string $alias alias of the table if used in the statement
+     * @return string
+     */
+    public function getDeletedWhere(string $alias = ''): string {
+        return $this->_REALDELETE ? '' : sprintf(' and %2$s%1$s=0 ', $this->_DELETEDFIELDNAME
+                        , !empty($alias) ? $this->_database->escapeFieldName($alias) . '.' : ''
+        );
     }
 
-    public function getFieldByAttribute($attributeName) {
-        if (!isset($this->_attributeToField[$attributeName])) {
+    /**
+     * Return field name by attribute to a data service
+     * @param string $attributeName
+     * @return string
+     */
+    public function getFieldByAttribute(string $attributeName): string {
+        if (empty($attributeName) || !isset($this->_attributeToField[$attributeName])) {
             return null;
         }
         return $this->_attributeToField[$attributeName];
     }
-    
-    protected function clearData() {
+
+    /**
+     * Clears the current object data
+     */
+    public function clearData() {
         $this->_data = array();
         reset($this->_dataFields);
         foreach ($this->_dataFields as $attributeData) {
@@ -112,23 +134,46 @@ abstract class CRUDObject {
         }
     }
 
-    protected function getEscapedTableKeyField() {
+    /**
+     * Returns the escaped name of the index field of the table
+     * @return string
+     */
+    protected function getEscapedTableKeyField(): string {
         return $this->_database->escapeFieldName($this->_tableKeyAttribute[1]);
     }
-    
-    public function getTableKeyField() {
+
+    /**
+     * Returns the name of the index field of the table
+     * @return string
+     */
+    public function getTableKeyField(): string {
         return $this->_tableKeyAttribute[1];
     }
 
+    /**
+     * Returns the escaped value of the current table key(primary index field)
+     * @return mixed
+     */
     protected function getEscapedTableKeyValue() {
         return $this->getEscapedAttributeValue($this->_tableKeyAttribute[0]);
     }
 
+    /**
+     * Returns the value of the current table key(primary index field)
+     * @return mixed
+     */
     protected function getTableKeyValue() {
         return $this->_data[$this->_tableKeyAttribute[0]];
     }
 
-    protected function getEscapedAttributeValue($attributeName) {
+    /**
+     * returns the escaped value of and attribute.
+     * 
+     * @param type $attributeName
+     * @return string
+     * @throws Exception
+     */
+    protected function getEscapedAttributeValue($attributeName): string {
 
         if (!array_key_exists($attributeName, $this->_data)) {
             throw new \Exception('Attempt to read a non existing attribute');
@@ -170,7 +215,14 @@ abstract class CRUDObject {
         return $data;
     }
 
-    protected function getEscapedValueByAttribute($attributeName, $attributeValue) {
+    /**
+     * Returns an escaped value based on attribute specification
+     * @param type $attributeName
+     * @param type $attributeValue
+     * @return string
+     * @throws Exception
+     */
+    protected function getEscapedValueByAttribute($attributeName, $attributeValue): string {
 
         if (!array_key_exists($attributeName, $this->_data)) {
             throw new \Exception('Attempt to read a non existing attribute');
@@ -211,6 +263,11 @@ abstract class CRUDObject {
         return $data;
     }
 
+    /**
+     * Fills the object with data by id(primary key value) or throws an exception
+     * @param mixed $id
+     * @throws Exception
+     */
     public function read($id = null) {
         $this->clearData();
 
@@ -221,8 +278,8 @@ abstract class CRUDObject {
                     , $this->getEscapedValueByAttribute($this->_tableKeyAttribute[0], $id)
                     , !$this->_REALDELETE ? sprintf(' and %s=0 ', $this->_database->escapeFieldName($this->_DELETEDFIELDNAME)) : '');
             $result = $this->_database->query($sql);
-            $foundData= 0;
-                    
+            $foundData = 0;
+
             foreach ($result as $data) {
                 $this->_data = array();
                 reset($this->_dataFields);
@@ -234,15 +291,15 @@ abstract class CRUDObject {
                         $this->_data[$attributeName] = null;
                     }
                 }
-                
-                $foundData=1;
+
+                $foundData = 1;
                 break;
             }
 
             if (!$foundData) {
-                throw new Exception('Item '.$id.' not found');
+                throw new Exception('Item ' . $id . ' not found');
             }
-            
+
             reset($this->_nestedObjects);
             foreach ($this->_nestedObjects as $attributeName => $className) {
                 if (is_string($className)) {
@@ -254,6 +311,10 @@ abstract class CRUDObject {
         }
     }
 
+    /**
+     * Save the record in the database
+     * @throws Exception
+     */
     public function save() {
         $sqlFields = array();
         $sqlData = array();
@@ -269,24 +330,24 @@ abstract class CRUDObject {
         try {
             reset($this->_dataFields);
             foreach ($this->_dataFields as $fieldName => $attributeData) {
-                if (empty($attributeData['canNotChange']) || empty($id)||!empty($attributeData['isPrimary'])) {
+                if (empty($attributeData['canNotChange']) || empty($id) || !empty($attributeData['isPrimary'])) {
                     if (empty($attributeData['isPrimary'])) {
-                        $field=$this->_database->escapeFieldName($fieldName);
+                        $field = $this->_database->escapeFieldName($fieldName);
                         $sqlFields[] = $field;
                         $val = $this->getEscapedAttributeValue($attributeData['attribute']);
                         $sqlData[] = $val;
-                        $updateData[] = $field.'=values('.$field.')';
+                        $updateData[] = $field . '=values(' . $field . ')';
                     } else {
                         //primary key is considered to not allow null
                         if (!is_object($this->_data[$attributeData['attribute']]) && $this->_data[$attributeData['attribute']] !== null) {
-                            $field=$this->_database->escapeFieldName($fieldName);
+                            $field = $this->_database->escapeFieldName($fieldName);
                             $sqlFields[] = $field;
                             $val = $this->getEscapedAttributeValue($attributeData['attribute']);
                             $sqlData[] = $val;
-                            $updateData[] = $field.'=values('.$field.')';
+                            $updateData[] = $field . '=values(' . $field . ')';
                         }
                     }
-                } 
+                }
             }
 
             $sql = sprintf('insert into %1$s (%2$s) values (%3$s) on duplicate key update %4$s'
@@ -319,6 +380,11 @@ abstract class CRUDObject {
         }
     }
 
+    /**
+     * Used to set id if needed
+     * @param type $id
+     * @throws Exception
+     */
     public function setId($id) {
         if (empty($this->_data[$this->_tableKeyAttribute[0]])) {
             $this->_data[$this->_tableKeyAttribute[0]] = $id;
@@ -327,10 +393,18 @@ abstract class CRUDObject {
         }
     }
 
+    /**
+     * Returns the id(primary key attribute) of the record
+     * @return type
+     */
     public function getId() {
         return $this->_data[$this->_tableKeyAttribute[0]];
     }
 
+    /**
+     * Full validation of the object data. Calls all validators and returns if they have failed or not.
+     * @return bool
+     */
     public function validateData(): bool {
         $result = true;
 
@@ -351,6 +425,12 @@ abstract class CRUDObject {
         return $result;
     }
 
+    /**
+     * Validate a single attribute of the data object calling all validators for the attribute.
+     * @param string $attributeName
+     * @param type $value
+     * @return bool
+     */
     protected function validateAttribute(string $attributeName, $value): bool {
         if (isset($this->_validators[$attributeName])) {
             if (is_array($this->_validators[$attributeName])) {
@@ -364,24 +444,29 @@ abstract class CRUDObject {
         }
     }
 
+    /**
+     * delete the record from the database;
+     */
     public function delete() {
         $id = $this->getTableKeyValue();
         $fieldName = $this->getEscapedTableKeyField();
-        if (!$this->_REALDELETE) {
-            $sql = sprintf('update %1$s set %2$s=1 where %3$s=%4$s'
-                    , $this->_database->escapeTableName($this->_TABLENAME)
-                    , $this->_database->escapeFieldName($this->_DELETEDFIELDNAME)
-                    , $this->getEscapedTableKeyField()
-                    , $this->getEscapedTableKeyValue()
-            );
-            $this->_database->execute($sql);
-        } else {
-            $sql = sprintf('delete from %1$s where %2$s=%3$s'
-                    , $this->_database->escapeTableName($this->_TABLENAME)
-                    , $this->getEscapedTableKeyField()
-                    , $this->getEscapedTableKeyValue()
-            );
-            $this->_database->execute($sql);
+        if (!empty($id)) {
+            if (!$this->_REALDELETE) {
+                $sql = sprintf('update %1$s set %2$s=1 where %3$s=%4$s'
+                        , $this->_database->escapeTableName($this->_TABLENAME)
+                        , $this->_database->escapeFieldName($this->_DELETEDFIELDNAME)
+                        , $this->getEscapedTableKeyField()
+                        , $this->getEscapedTableKeyValue()
+                );
+                $this->_database->execute($sql);
+            } else {
+                $sql = sprintf('delete from %1$s where %2$s=%3$s'
+                        , $this->_database->escapeTableName($this->_TABLENAME)
+                        , $this->getEscapedTableKeyField()
+                        , $this->getEscapedTableKeyValue()
+                );
+                $this->_database->execute($sql);
+            }
         }
     }
 
@@ -405,15 +490,21 @@ abstract class CRUDObject {
         if (isset($this->_readOnlyAttributes[$name]) && !empty($this->getTableKeyValue()) || $this->_tableKeyAttribute[0] == $name) {
             throw new \Exception('Can not change read only attributes (' . $name . ')');
         }
-        
+
         if (isset($this->_setterPreprocessors[$name])) {
-            $value=$this->{$this->_setterPreprocessors[$name]}($value);
+            $value = $this->{$this->_setterPreprocessors[$name]}($value);
         }
 
         $this->_data[$name] = $value;
     }
 
+    /**
+     * Validates a field that should not be empty
+     * @param type $value
+     * @return bool
+     */
     protected function notEmptyValidator($value): bool {
         return !empty($value);
     }
+
 }
